refacto(Inventory): rename methods
