+++ src/app/inventory/components/Inventory/index.js
@@ -7,12 +7,12 @@ import selectSelectedItemId from "../../redux/reducers/sceneState/selectors/sele
 import { initModule } from "../../redux/actions/init";
 import { setCurrentContext, CONTEXT_TYPE } from "../../redux/actions/inventory";
 import { updateItemsInventoriesRequest } from "../../redux/actions/inventories";
-import getItemsIdsFromInventoryByCategorySelected from "../../redux/reducers/entities/inventories/getters/get-items-ids-from-inventory-by-category-selected";
+import getItemsIdsFromCurrentInventoryByCategorySelected from "../../redux/reducers/entities/inventories/getters/get-items-ids-from-current-inventory-by-category-selected";
 import getSelectedCategory from "../../redux/reducers/sceneState/getters/get-selected-category";
 
 const mapStateToProps = state => {
   const selectedItemId = selectSelectedItemId(state);
-  const itemsIds = getItemsIdsFromInventoryByCategorySelected(state);
+  const itemsIds = getItemsIdsFromCurrentInventoryByCategorySelected(state);
   const { name } = getSelectedCategory(state);
   return {
     selectedItemId,
