+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-categories-from-current-inventory.js
