+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-has-items-form-current-inventory.js
