+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-item-completed-from-current-inventory-by-id.js
