+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-items-completed-from-current-inventory.js
